[Refactoring] Use animeGameDataModels for weather rersources

+++ src/main/java/emu/grasscutter/game/world/Scene.java
@@ -40,6 +40,7 @@ import lombok.Setter;
 import lombok.val;
 import org.anime_game_servers.game_data_models.gi.data.dungeon.DungeonData;
 import org.anime_game_servers.game_data_models.gi.data.dungeon.DungeonPassConditionType;
+import org.anime_game_servers.game_data_models.gi.data.scene.weather.ClimateType;
 import org.anime_game_servers.gi_lua.models.SceneGroupUserData;
 import org.anime_game_servers.gi_lua.models.ScriptArgs;
 import org.anime_game_servers.gi_lua.models.constants.EventType;
@@ -1156,7 +1157,7 @@ public class Scene {
 
             if(w != null && !this.weatherAreas.containsKey(e.getKey())) {
                 WeatherArea area = new WeatherArea(this, w);
-                area.setClimateType(ClimateType.getTypeByValue((int)e.getValue()));
+                area.setClimateType(ClimateType.getTypeByValue(e.getValue()));
                 weatherAreas.put((int)e.getKey(), area);
             }
         });
